competitor_analysis/requirements.txt
# ============================================================
# AI Startup Investment Evaluation Agent - Requirements
# ============================================================

# Core Dependencies
# ------------------------------------------------------------
python-dotenv==1.0.0

# LangChain Ecosystem
# ------------------------------------------------------------
langchain==0.3.13
langchain-core==0.3.28
langchain-community==0.3.13
langchain-openai==0.2.14
langchain-text-splitters==0.3.4

# LangGraph (Multi-Agent Framework)
# ------------------------------------------------------------
langgraph==0.2.57

# LangChain Teddynote (Custom Utilities)
# ------------------------------------------------------------
langchain-teddynote==0.3.5

# Vector Store & Embeddings
# ------------------------------------------------------------
faiss-cpu==1.9.0
# faiss-gpu==1.9.0  # GPU 버전 사용 시 주석 해제

# Document Loaders
# ------------------------------------------------------------
pypdf==5.1.0

# Data Validation & Schema
# ------------------------------------------------------------
pydantic==2.10.4
pydantic-core==2.27.2
typing-extensions==4.12.2

# Web Search (Tavily API)
# ------------------------------------------------------------
tavily-python==0.5.0

# OpenAI API
# ------------------------------------------------------------
openai==1.59.7

# Additional Utilities
# ------------------------------------------------------------
requests==2.32.3
httpx==0.28.1

# Optional: Development & Testing
# ------------------------------------------------------------
# jupyter==1.1.1
# ipykernel==6.29.5
# pytest==8.3.4
# black==24.10.0
